Add sequential variational inference

+++ dataio/path.py
@@ -1,12 +1,18 @@
 """Path module."""
 
+from pathlib import Path
 
-def rmtree(path):
-    """Remove directory."""
+def clear(path):
+    """Clear files and directories under a directory."""
+    path = Path(path)
     if path.exists():
+        if not path.is_dir():
+            raise ValueError('path is not of a directory')
         for p in path.iterdir():
             if p.is_file():
                 p.unlink()
             else:
-                rmtree(p)
-        path.rmdir()
+                clear(p)
+                p.rmdir()
+    else:
+        path.mkdir(parents=True)
